requirements.txt
# PyTorch with ROCm 5.6 support (AMD GPU)
torch==2.2.0+rocm5.6
torchvision==0.17.0+rocm5.6
torchaudio==2.2.0+rocm5.6
-f https://download.pytorch.org/whl/rocm5.6/torch_stable.html

# Ultralytics YOLOv8
ultralytics==8.2.21

roboflow==1.1.29

# Other useful packages
PyYAML>=5.3.1
